int->float

+++ GOTO_EngineLib/inc/D2DRenderAPI.h
@@ -34,8 +34,8 @@ namespace GOTOEngine
         void Clear() override;
         //void DrawImage(int x, int y, float scale, bool flipX, const IRenderImage* image) override;
         void DrawBitmap(const Matrix3x3& mat, IRenderBitmap* bitmap) override;
-        void DrawString(int x, int y, int width, int height, const wchar_t* string, const IRenderFont* font, bool rightAlign, Color color) override;
-        void DrawRect(int x, int y, int width, int height, bool fill, Color color) override;
+        void DrawString(float x, float y, float width, float height, const wchar_t* string, const IRenderFont* font, bool rightAlign, Color color) override;
+        void DrawRect(float x, float y, float width, float height, bool fill, Color color) override;
 
         void SetViewport(Rect rect) override;
         void ResetViewport() override;
